+++ src/debug/Data.jl
@@ -29,7 +29,7 @@ const strategies = let
     ),
   )
 
-  light = merge(minimalistic, Dict("StrategyLevel" => 1))
+  light = merge(minimalistic, Dict("AddGitHubPRTemplate" => true, "StrategyLevel" => 1))
 
   moderate =
     merge(light, Dict("AddCitationCFF" => true, "AddDependabot" => true, "StrategyLevel" => 2))