+++ src/debug/Data.jl
@@ -31,7 +31,8 @@ const strategies = let
 
   light = merge(minimalistic, Dict("StrategyLevel" => 1))
 
-  moderate = merge(light, Dict("AddCitationCFF" => true, "StrategyLevel" => 2))
+  moderate =
+    merge(light, Dict("AddCitationCFF" => true, "AddDependabot" => true, "StrategyLevel" => 2))
 
   robust = merge(
     moderate,