+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning].
 - New question: `AddCitationCFF` to make the CITATION.CFF optional.
 - New question: `AddDependabot` to make `.github/dependabot.yml` optional.
 - New question: `AddGitHubPRTemplate` to make `.github/PULL_REQUEST_TEMPLATE.md` optional.
+- New question: `AddLintCI` to make `.github/workflows/Lint.yml` optional.
 
 ### Changed