+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning].
 - New question: `AddCompatHelperCI` to make `.github/workflows/CompatHelper.yml` optional.
 - New question: `AddTagBotCI` to make `.github/workflows/TagBot.yml` optional.
 - New question: `AddTestCI` to make `.github/workflows/`'s `Test.yml`, `TestOnPRs`, and `ReusableTest.yml` optional.
+- New question: `AddLychee` to make `.lychee.toml` optional.
 
 ### Changed